+++ packages/app/actions/user.test.tsx
@@ -7,20 +7,75 @@ describe('reducer', () => {
   });
 
   it('should handle LOGIN_SUCCESS actions', () => {
-    const result = reducer(initialState, { type: 'user/LOGIN_SUCCESS', user: { id: 1234 } });
-    expect(result).toStrictEqual({ initialized: false, user: { id: 1234 } });
+    const result = reducer(initialState, {
+      type: 'user/LOGIN_SUCCESS',
+      user: {
+        sub: '1234',
+        email: null,
+        scope: 'apps:read',
+        name: null,
+        email_verified: false,
+        picture: null,
+      },
+      role: null,
+    });
+    expect(result).toStrictEqual({
+      initialized: false,
+      user: {
+        sub: '1234',
+        email: null,
+        scope: 'apps:read',
+        name: null,
+        email_verified: false,
+        picture: null,
+      },
+      role: null,
+    });
   });
 
   it('should handle LOGOUT actions', () => {
-    const result = reducer({ initialized: true, user: { id: 1234 } }, { type: 'user/LOGOUT' });
-    expect(result).toStrictEqual({ initialized: true, user: null });
+    const result = reducer(
+      {
+        initialized: true,
+        user: {
+          sub: '1234',
+          email: null,
+          scope: 'apps:read',
+          name: null,
+          email_verified: false,
+          picture: null,
+        },
+        role: 'Test',
+      },
+      { type: 'user/LOGOUT' },
+    );
+    expect(result).toStrictEqual({ initialized: true, user: null, role: null });
   });
 
   it('should handle INITIALIZED actions', () => {
-    const result = reducer(initialState, { type: 'user/INITIALIZED', user: { id: 1234 } });
+    const result = reducer(initialState, {
+      type: 'user/INITIALIZED',
+      user: {
+        sub: '1234',
+        email: null,
+        scope: 'apps:read',
+        name: null,
+        email_verified: false,
+        picture: null,
+      },
+      role: null,
+    });
     expect(result).toStrictEqual({
       initialized: true,
-      user: { id: 1234 },
+      user: {
+        sub: '1234',
+        email: null,
+        scope: 'apps:read',
+        name: null,
+        email_verified: false,
+        picture: null,
+      },
+      role: null,
     });
   });
 });