+++ packages/app/components/BlockList/index.tsx
@@ -1,1 +1,13 @@
-export { default } from './BlockList';
+import { connect } from 'react-redux';
+
+import { State } from '../../actions';
+import BlockList from './BlockList';
+
+function mapStateToProps(state: State): Partial<React.ComponentPropsWithoutRef<typeof BlockList>> {
+  return {
+    security: state.app.definition.security,
+    role: state.user.role,
+  };
+}
+
+export default connect(mapStateToProps)(BlockList);