+++ packages/app/components/ProfileDropdown/messages.tsx
@@ -1,9 +1,8 @@
 import { defineMessages } from 'react-intl';
 
 export default defineMessages({
-  documentation: 'Documentation',
   logoutButton: 'Logout',
-  login: 'Login',
   settings: 'Settings',
   pfp: 'Profile Picture',
+  login: 'Login',
 });