+++ packages/server/controllers/apps.js
@@ -1,7 +1,9 @@
 import { logger } from '@appsemble/node-utils';
 import { normalize, permissions, StyleValidationError, validateStyle } from '@appsemble/utils';
+import validateAppDefinition, {
+  AppsembleValidationError,
+} from '@appsemble/utils/validateAppDefinition';
 import Boom from '@hapi/boom';
-import Ajv from 'ajv';
 import crypto from 'crypto';
 import jsYaml from 'js-yaml';
 import { isEqual, uniqWith } from 'lodash';
@@ -10,62 +12,26 @@ import sharp from 'sharp';
 import * as webpush from 'web-push';
 
 import checkRole from '../utils/checkRole';
-import getAppBlocks from '../utils/getAppBlocks';
 import getAppFromRecord from '../utils/getAppFromRecord';
 import getDefaultIcon from '../utils/getDefaultIcon';
 
-const ajv = new Ajv();
-ajv.addFormat('fontawesome', () => true);
+function getBlockVersions(db) {
+  return async blocks => {
+    const blockVersions = await db.models.BlockVersion.findAll({
+      raw: true,
+      where: {
+        [Op.or]: uniqWith(
+          Object.values(blocks).map(({ type, version }) => ({
+            name: type.startsWith('@') ? type : `@appsemble/${type}`,
+            version,
+          })),
+          isEqual,
+        ),
+      },
+    });
 
-async function checkBlocks(definition, db) {
-  const blocks = getAppBlocks(definition);
-  const blockVersions = await db.models.BlockVersion.findAll({
-    raw: true,
-    where: {
-      [Op.or]: uniqWith(
-        Object.values(blocks).map(({ type, version }) => ({
-          name: type.startsWith('@') ? type : `@appsemble/${type}`,
-          version,
-        })),
-        isEqual,
-      ),
-    },
-  });
-  const blockVersionMap = new Map();
-  blockVersions.forEach(version => {
-    if (!blockVersionMap.has(version.name)) {
-      blockVersionMap.set(version.name, new Map());
-    }
-    blockVersionMap.get(version.name).set(version.version, version);
-  });
-  const errors = Object.entries(blocks).reduce((acc, [loc, block]) => {
-    const type = block.type.startsWith('@') ? block.type : `@appsemble/${block.type}`;
-    const versions = blockVersionMap.get(type);
-    if (!versions) {
-      return { ...acc, [loc]: `Unknown block type “${type}”` };
-    }
-    if (!versions.has(block.version)) {
-      return { ...acc, [loc]: `Unknown block version “${type}@${block.version}”` };
-    }
-    const version = versions.get(block.version);
-    if (Object.prototype.hasOwnProperty.call(version, 'parameters')) {
-      const validate = ajv.compile(version.parameters);
-      const valid = validate(block.parameters);
-      if (!valid) {
-        return validate.errors.reduce(
-          (accumulator, error) => ({
-            ...accumulator,
-            [`${loc}.parameters${error.dataPath}`]: error,
-          }),
-          acc,
-        );
-      }
-    }
-    return acc;
-  }, null);
-  if (errors) {
-    throw Boom.badRequest('Block validation failed', errors);
-  }
+    return blockVersions;
+  };
 }
 
 function handleAppValidationError(error, app) {
@@ -73,6 +39,10 @@ function handleAppValidationError(error, app) {
     throw Boom.conflict(
       `Another app with path “@${app.OrganizationId}/${app.path}” already exists`,
     );
+  }
+
+  if (error instanceof AppsembleValidationError) {
+    throw Boom.badRequest('Appsemble definition is invalid.', error.data || error.message);
   }
 
   if (error instanceof StyleValidationError) {
@@ -135,7 +105,7 @@ export async function createApp(ctx) {
     }
 
     await checkRole(ctx, OrganizationId, permissions.CreateApps);
-    await checkBlocks(definition, db);
+    await validateAppDefinition(definition, getBlockVersions(ctx.db));
 
     for (let i = 1; i < 11; i += 1) {
       const p = i === 1 ? path : `${path}-${i}`;
@@ -267,7 +237,7 @@ export async function updateApp(ctx) {
       result.yaml = jsYaml.safeDump(definition);
     }
 
-    await checkBlocks(definition, db);
+    await validateAppDefinition(definition, getBlockVersions(ctx.db));
 
     const dbApp = await App.findOne({ where: { id: appId } });
 
@@ -307,7 +277,7 @@ export async function patchApp(ctx) {
 
     if (definition) {
       result.definition = definition;
-      await checkBlocks(definition, db);
+      await validateAppDefinition(definition, getBlockVersions(ctx.db));
     }
 
     if (path) {
@@ -402,6 +372,127 @@ export async function deleteApp(ctx) {
 
   await app.update({ path: null });
   await app.destroy();
+}
+
+export async function getAppMembers(ctx) {
+  const { appId } = ctx.params;
+  const { App, User } = ctx.db.models;
+
+  const app = await App.findByPk(appId, { include: [User] });
+  if (!app) {
+    throw Boom.notFound('App not found');
+  }
+
+  ctx.body = app.Users.map(user => ({
+    id: user.id,
+    name: user.name,
+    primaryEmail: user.primaryEmail,
+    role: user.AppMember.role,
+  }));
+}
+
+export async function getAppMember(ctx) {
+  const { appId, memberId } = ctx.params;
+  const { App, User, Organization } = ctx.db.models;
+
+  const app = await App.findByPk(appId, {
+    include: [{ model: User, where: { id: memberId }, required: false }, Organization],
+  });
+  if (!app) {
+    throw Boom.notFound('App not found');
+  }
+
+  if (app.definition.security === undefined) {
+    throw Boom.notFound('App does not have a security definition.');
+  }
+
+  const { policy, role: defaultRole } = app.definition.security.default;
+
+  const user = await User.findByPk(memberId);
+
+  if (!user) {
+    throw Boom.notFound('User does not exist.');
+  }
+
+  const member = app.Users.find(u => u.id === memberId);
+  let role;
+
+  if (member) {
+    role = member.AppMember.role;
+  } else {
+    switch (policy) {
+      case 'everyone':
+        role = defaultRole;
+
+        break;
+
+      case 'organization':
+        if (!(await app.Organization.hasUser(memberId))) {
+          throw Boom.notFound('User is not a member of the organization.');
+        }
+
+        role = defaultRole;
+
+        break;
+
+      case 'invite':
+        throw Boom.notFound('User is not a member of the app.');
+
+      default:
+        role = null;
+    }
+  }
+
+  ctx.body = {
+    id: user.id,
+    name: user.name,
+    primaryEmail: user.primaryEmail,
+    role,
+  };
+}
+
+export async function setAppMember(ctx) {
+  const { appId, memberId } = ctx.params;
+  const { role } = ctx.request.body;
+  const { App, User } = ctx.db.models;
+
+  const app = await App.findByPk(appId, { include: [User] });
+  if (!app) {
+    throw Boom.notFound('App not found');
+  }
+
+  await checkRole(ctx, app.OrganizationId, permissions.EditApps);
+
+  const user = await User.findByPk(memberId);
+  if (!user) {
+    throw Boom.notFound('User with this ID doesn’t exist.');
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(app.definition.security.roles, role)) {
+    throw Boom.badRequest(`Role ‘${role}’ is not defined.`);
+  }
+
+  const [member] = await app.getUsers({ where: { id: memberId } });
+
+  if (member) {
+    if (
+      role === app.definition.security.default.role &&
+      app.definition.security.default.policy !== 'invite'
+    ) {
+      await app.removeUser(member);
+    } else {
+      await member.AppMember.update({ role });
+    }
+  } else {
+    await app.addUser(user, { through: { role } });
+  }
+
+  ctx.body = {
+    id: user.id,
+    name: user.name,
+    primaryEmail: user.primaryEmail,
+    role,
+  };
 }
 
 export async function getAppRatings(ctx) {