+++ packages/studio/components/AppSideMenu/AppSideMenu.tsx
@@ -26,7 +26,7 @@ export default function AppSideMenu({ app, match }: AppSideMenuProps): React.Rea
   const { userInfo } = useUser();
   const [isCollapsed, setCollapsed] = React.useState(false);
   const organizations = useOrganizations();
-  const organization = organizations.find(org => org.id === app.OrganizationId);
+  const organization = organizations && organizations.find(org => org.id === app.OrganizationId);
 
   return (
     <SideMenu isCollapsed={isCollapsed} toggleCollapse={() => setCollapsed(!isCollapsed)}>
@@ -89,6 +89,16 @@ export default function AppSideMenu({ app, match }: AppSideMenuProps): React.Rea
               </span>
             </NavLink>
           )}
+          {organization &&
+            app.definition.security !== undefined &&
+            checkRole(organization.role, permissions.EditApps) && (
+              <NavLink className={styles.menuItem} exact={!isCollapsed} to={`${match.url}/roles`}>
+                <Icon icon="users" size="medium" />
+                <span className={classNames({ 'is-hidden': isCollapsed })}>
+                  <FormattedMessage {...messages.roles} />
+                </span>
+              </NavLink>
+            )}
           {organization && checkRole(organization.role, permissions.EditAppSettings) && (
             <NavLink className={styles.menuItem} exact={!isCollapsed} to={`${match.url}/settings`}>
               <Icon icon="cogs" size="medium" />