+++ packages/studio/components/AppSideMenu/messages.tsx
@@ -6,4 +6,5 @@ export default defineMessages({
   resources: 'Resources',
   settings: 'Settings',
   notifications: 'Notifications',
+  roles: 'Roles',
 });