+++ packages/studio/types.ts
@@ -1,39 +1,4 @@
 /**
- * OpenID Connect specifies a set of standard claims about the end-user, which cover common profile
- * information such as name, contact details, date of birth and locale.
- *
- * The Connect2id server can be set up to provide additional custom claims, such as roles and
- * permissions.
- */
-export interface UserInfo {
-  /**
-   * The subject (end-user) identifier. This member is always present in a claims set.
-   */
-  sub: number;
-
-  /**
-   * The full name of the end-user, with optional language tag.
-   */
-  name: string;
-
-  /**
-   * The end-user's preferred email address.
-   */
-  email: string;
-
-  /**
-   * True if the end-user's email address has been verified, else false.
-   */
-  // eslint-disable-next-line camelcase
-  email_verified: boolean;
-
-  /**
-   * The URL of the profile picture for the end-user.
-   */
-  picture: string;
-}
-
-/**
  * A user email registration.
  */
 export interface UserEmail {